Rewrite io.c, including its API.
io.c almost complicated enough to be a work of absurdist humor.  This
version uses only one socket, with the most essential functions.  The
only ornament is getnameinfo(), so that the user can see the address and
hostname of the device actually being reflashed.

This revision also simplifies the input/output buffering by keeping a
local line pointer in the TCP handle for return strings in tcp_io(),
which were retrieved by getline().