Get line from device manually.
I kept getting ``illegal seek" errors.  This may be because the T680
replies a prompt *after* EOL, so getline() was not reading that in
before the next call to tcp_io*().  Perhaps for some libc implementations
``r+" doesn't create separate input and output buffers - I really can't
find any good documentation on this.

So for TCP input we just call read() instead of getc().  The bummer
is that I had to write my own getline implementation, which I
called ``sock_getline()".